+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Bugfixes:
 
 New features:
 - Add support for starting a repl within a folder which has not been setup as a spago project (#168)
+- Add `--format` flag to `spago docs` (#294)
 
 ## [0.8.5] - 2019-06-18